Save-tiny-components, picker, and pos normalize;
